Added support for LeapJS lib version and bug fix
You can now pass in the LeapJS version number in case you are bound to a specific version. Also put in place a fix for timeout when frame data hasn't been received within the expected amount of time, such as when Leap software is running but the device is not hooked up, and or not responding.